Closes #165, Working Aspect Ratio: - if slider = -1, only choose nodes with a.r. >= prev a.r. - else, correct a.r. to desired, then choose nodes that dont violate it

+++ src/util/DisplayMessagesGui.java
@@ -46,7 +46,7 @@ public class DisplayMessagesGui {
 
     public static String createAspectRatioMsg(IGraph graph) {
         GraphOperations.AspectRatio aspectRatio = GraphOperations.getAspectRatio(graph);
-        return "Aspect Ratio: " + (int)aspectRatio.getValue() +
+        return "Aspect Ratio: " + aspectRatio.getValue() +
                 " | Min: " + (int)aspectRatio.getShortestEdgeLength() +
                 " | Max.: " + (int)aspectRatio.getLongestEdgeLength();
     }
